Safari-611.1.10/LayoutTests/fast/xmlhttprequest/xmlhttprequest-sync-xhr-failure-loop-during-unload.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/js-test.js"></script>
<iframe id="testFrame" src="resources/xmlhttprequest-sync-xhr-failure-loop-during-unload-iframe.html"></iframe>
<script>
description("Makes sure that retrying failed sync XHRs in a loop during unload does not cause a hang.");
jsTestIsAsync = true;

frameDidUnload = () => {
    testPassed("No hang while unloading the iframe");
    finishJSTest();
}

onload = () => {
    testFrame.remove();
}
</script>
</html>
